+++ src/types/props.types.ts
@@ -56,4 +56,5 @@ export interface AuraProps {
     accentInsensitiveSearch?: boolean;
     highlightSearchResults?: boolean;
     highlightClass?: string | null;
+    cellClickSearch?: boolean;
 }